move formatting of fileName strings to FileNameFormatter
