+++ universal-application-tool-0.0.1/app/controllers/dev/AzuriteFileStorageStrategy.java
@@ -7,7 +7,7 @@ import models.StoredFile;
 import play.mvc.Http.Request;
 import play.mvc.Result;
 import repository.StoredFileRepository;
-import services.cloud.azure.FileNameFormatter;
+import services.cloud.FileNameFormatter;
 
 /** Implements file uploading to Azurite, the Azure emulator. */
 public class AzuriteFileStorageStrategy implements CloudEmulatorFileStorageStrategy {
