+++ universal-application-tool-0.0.1/test/services/cloud/FileNameFormatterTest.java
@@ -1,8 +1,9 @@
-package services.cloud.azure;
+package services.cloud;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import org.junit.Test;
+import services.cloud.FileNameFormatter;
 
 public class FileNameFormatterTest {
